requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0.0
rich>=13.0.0
